MasterUI changed back to StoryGame for familiarity's sake. EncounterView, PlayerCreationView, and MapView all are accessed from StoryGame for a better interface flow.
