+++ src/main/java/edu/bsu/storygame/StoryGame.java
@@ -1,14 +1,12 @@
 package edu.bsu.storygame;
 
+import edu.bsu.storygame.views.MapView;
+import edu.bsu.storygame.views.PlayerCreationView;
 import javafx.application.Application;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
 import react.Slot;
+
 
 public class StoryGame extends Application {
 
@@ -16,64 +14,41 @@ public class StoryGame extends Application {
         launch(args);
     }
 
+    final GameContext context = new GameContext();
+
     @Override
     public void start(final Stage primaryStage) throws Exception {
-        final GameContext context = new GameContext();
-        PhaseLabel phaseLabel = new PhaseLabel(context);
-        Button button = new Button("Change phase!");
-        final MapView mapScreen = new MapView(context);
-        final Stage mapStage = mapScreen.mapStage();
-        mapStage.setTitle("Nightmare Defenders Map");
-        button.setOnAction(new EventHandler<ActionEvent>() {
-            public void handle(ActionEvent event) {
-                if (context.phase.get().equals(Phase.MOVEMENT)) {
-                    context.phase.update(Phase.ENCOUNTER);
-                    mapScreen.africaRegion.setDisable(true);
-                    mapScreen.europeRegion.setDisable(true);
-                } else {
-                    context.phase.update(Phase.MOVEMENT);
-                    mapScreen.africaRegion.setDisable(false);
-                    mapScreen.europeRegion.setDisable(false);
+        context.phase.update(Phase.PLAYER_CREATION);
+        primaryStage.setScene(setPlayerCreationScene());
+        context.phase.connect(new Slot<Phase>() {
+            @Override
+            public void onEmit(Phase phase) {
+                if(context.phase.get().equals(Phase.PLAYER_CREATION)){
+                    primaryStage.setScene(setPlayerCreationScene());
+                }
+                if(context.phase.get().equals(Phase.MOVEMENT)){
+                    primaryStage.setScene(setMapViewScene());
+                }
+                if(context.phase.get().equals(Phase.ENCOUNTER)){
+                    showEncounterView();
                 }
             }
+
         });
 
-
-        Button loadMapViewButton = new Button("Go To Game Map!");
-        loadMapViewButton.setOnAction(new EventHandler<ActionEvent>() {
-            public void handle(ActionEvent event) {
-                try {
-                    mapStage.show();
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
-            }
-        });
-
-        primaryStage.setTitle("Spring Studio Project");
-        VBox root = new VBox();
-        root.getChildren().add(phaseLabel);
-        root.getChildren().add(button);
-        root.getChildren().add(loadMapViewButton);
-        primaryStage.setScene(new Scene(root));
         primaryStage.show();
-
+    }
+    private Scene setPlayerCreationScene(){
+        PlayerCreationView view = new PlayerCreationView(context);
+        return view.getPlayerCreationScene();
     }
 
-    private class PhaseLabel extends Label {
-
-        public PhaseLabel(GameContext context) {
-            super(context.phase.get().toString());
-            context.phase.connect(new Slot<Phase>() {
-                @Override
-                public void onEmit(Phase phase) {
-                    PhaseLabel.this.setText(phase.toString());
-                }
-            });
-        }
+    private Scene setMapViewScene() {
+        MapView view = new MapView(context);
+        return view.initMap();
     }
 
-
-
-
+    private void showEncounterView(){
+        new WraithEncounter(context).show();
+    }
 }
